Compute network flows
This commit computes flows from a packet stream.  It is based on
the standard 5-tuple.  Flows terminate under four conditions:
active timeout, idle timeout, normal session termination, and
resource exhaustion.

Two debug options are provided for the command line. Each is turned
off by default.

 `--debug-print-packets` will print each parsed (meta)packet in
        short form

 `--debug-print-flows` will print each flow in short form`

Flows are written to files, which are rotated when the file size
reaches 100MB. We currently use the lumberjack package but should
develop a custom flow log package with finer grain rotation logic.
The output prefix is currently set to `./output` and the flow file
has a prefix slug of `flow-`.

TODO

This is a very Version 0.1.  There is much work left to be done. The
following are the two big remaining issues.

1. Refactor and redesign using a pipeline workflow

2. Handle idle timeouts more efficiently.

NOTE: The idle timeout cache is a *very* unoptimized implementation.
It is a simple get-it-done approach using slices with very terrible
performance due to the massive amount slice resizing.
*DO NOT USE THIS VERSION BEYOND THE EARLY PROTOTYPE STAGE*

A potentially better approach is a custom linked list that implicitly
stores pointers with the elements, i.e. {next, hash, expiry}, with
element allocations from a memory pool of size capacity. Go's
`container/list` is not an option because elements are "external" to
the list and use a "Value interface" to access the element.  Heap
allocations (and assocaited GC activity) and general interfaces are
not really our use case for an IdleCache.

Some other ideas are to borrow from
 [golang-lru](https://github.com/hashicorp/golang-lru)
 [freecache](https://github.com/coocood/freecache)
 [go-cache](https://godoc.org/github.com/robfig/go-cache)
 [Memory pools using channels](http://stackoverflow.com/questions/38505830/how-to-implement-memory-pooling-in-golang)