Create MetaPacket
Until we figure out a way to better manage zero-copy across channels,
this commit provides a MetaPacket struct to capture packet data we
will use for flows.  We will need to captures some payload bytes later
on for banners.

The main insight for this commit is that the composite data structures
in gopacket contain slices.  They don't copy well across channels because
slices are copy-by-value, and that doesn't copy the underlying array.
The source and destination addresses are slices, hence we need to use
something like MetaPacket.

This may be an optimization target.  Performance benchmarking is required
for this design.