Flow cache
We provide a flow cache for aggregating and terminating flows that is
very much more performant than the previous approach (an uncoupled hash-
map and slices).  Our FlowCache is a combination of a hashmap and a
pick queue, which allows us to efficiently implement idle timeouts,
active timeouts, and normal terminations.

A pick queue allows removal of nodes from the middle of the queue, not
just the top (strictly speaking, we also allow push and pops from both
head and tail). Each node has a key into the hashmap and an idle expiry.
So, during idle timeout checks, we quickly go through the list and purge
from the hash map. Otherwise, hashmap elements have a pointer to a
corresponding pick queue element, which permits O(1) removal, i.e. not
queue traversal.  So, flow access, random removal, and idle timeouts are
all O(1).

Yeah, it's much faster.